fix: use correct args and kwargs definitions
Various methods are defined with improper vardadic positional and keyword-arguments. The original intention for @overload signatures was to provide type support for passthroughs to API query parameter or json arguments. Along the way, *args was introduced to these methods assuming that we wanted to match the dict function signature. In practice, this is confusing since passing SupportsKeysAndGetItem typed variables is rare. Threfore, the *args part of the function signature is removed wherever possible. In the case of update(self, *args, **kwargs) we cannot remove *args since we inherit from dict and dict enforces this method signature.