build: always use uv for dependency management (#265)
Modifies the build tooling to always use `uv pip` instead of `pip` and `uv build` instead of `python -m build`.

Additional checks are added to ensure that `uv` exists on the system. If
`uv` does not exists, then it is installed via `pip`. If `pip` is not
installed, it is installed via `python -m ensurepip`. The `pip` check is
required for a virtual environment created by `uv venv` since by
default, `uv venv` does not install `pip`.

The UV_SYSTEM_PYTHON environment variable is set in CI and Docker to
force `uv` to use the system Python environment instead of depending on
a virtual environment. See
https://docs.astral.sh/uv/configuration/environment/

This reduces the execution time of `make` from 22s to 5s on my machine
without a Pip cache. With a hydrated PIp cache, the execution time is
reduced from 8s to 4s.