feat: add vanities (#310)
Adds vanities support.

A new mixin pattern is introduced, which adds the `vanity` attributes to
the `ContentItem` class. This abstraction is still a bit leaky since the
endpoint is hardcoded to 'v1/content' . But I find this design helpful
since it collocates everything regarding vanities into a single file. In
addition, the Content class has started to get a little unwieldy due to
its size.

The use of TypedDict, Unpack, Required, and NotRequired from
`typing_extensions` is also introduced. Thanks to
https://blog.changs.co.uk/typeddicts-are-better-than-you-think.html, I
discovered this pattern provides a much more succinct way to define
pass-through arguments. It also greatly simplifies the verbosity of
overload methods. That isn't shown here, but I have a test on another
branch.

Resolves #175