feat: add jobs (#314)
Adds job support.

The mixin pattern is continued, and a 'jobs' property is added to
`ContentItem.`

A new abstraction around resources called `Active` is introduced. This
abstraction begins transitioning to a more idiomatic approach to
interacting with collections. Similar to how printing a `Resource` shows
a `dict`, the `ActiveSequence` shows a familiar `List[dict]`
representation. The `ActiveFinderMethods` supports the `find` and
`find_by` methods. This is inspired by this Ruby on Rails module
https://api.rubyonrails.org/classes/ActiveRecord/FinderMethods.html.
Ideally, this separation of concerns provides a straightforward way to
compose bindings in the future and reduces the amount of duplication
across implementations.

Additional background may be found here
https://github.com/tdstein/active

Related to #310 
Resolves #306