feat: Add client credential oauth integration support + related databricks helpers to SDK (#348)
This PR adds a new function to the oauth tooling. It knows how to craft
a request against the credential exchange endpoint that can interface
with client credential flows. This supports service-account-like oauth
interactions.

Connect sets an environment variable called
`CONNECT_CONTENT_SESSION_TOKEN` which can be used by the executing
content when making the request. This means that the syntax for using
this new credential exchange function is slightly different.

This PR also adds the databricks extensions needed to use this new oauth
exchange flow when interfacing with databricks resources.

This feature won't land in Connect until the upcoming release gets
published - I don't know how we typically queue up work like this in
sync with product releases.

---------

Co-authored-by: Barret Schloerke <barret@posit.co>