feat: update Databricks helpers to be more explicit
Databricks sdk helpers now include a helper `databricks_config()`
which allows for more fine grained control over how the
credentials_strategy is selected in content that needs to run in
multiple environments (Workbench, Connect, local laptop). We now do our
best to "discover" the user's environment before selecting a strategy.

If no strategy was provided for the current environment then we use the
Databricks SDK's `DefaultCredentials()` class as a fallback option.

This is a breaking change but should simplify the API considerably for
users.