feat: speed up the packages api
The Connect Packages API (v1/packages) supports uncapped page size. By increasing the page size from the default of 500 to 1,000,000, there is a significant speed up when fetching all packages. This is due to a reduction in the number of round trips to the Connect server. The tradeoff is a larger payload. Since the Python requests package does not have a HTTP GET request-response body size limit, we are not concerned with the larger payload.