feat: speed up the packages api (#396)
The Connect Packages API (v1/packages) supports uncapped page size. By
increasing the page size from the default of 500 to 1,000,000, there is
a significant speed up when fetching all packages. This is due to a
reduction in the number of round trips to the Connect server. The
tradeoff is a larger payload. Since the Python requests package does not
limit the HTTP GET request-response body size, we are unconcerned with
the larger payload.